requirements.txt
Flask==2.2.2
numpy==1.23.5
Pillow==9.4.0
Pillow==10.0.0
ultralytics==8.0.159
Werkzeug==2.2.2
